feat(utils): update db exports to include takeOne and takeUniqueOrThrow functions
