+++ apps/api/src/plugins/account.ts
@@ -1,6 +1,6 @@
+import { db } from "@ponti/utils";
 import { account } from "@ponti/utils/schema";
 import { eq } from "drizzle-orm";
-import { db } from "../db";
 
 export function getUserAccount(userId: string) {
 	return db.select().from(account).where(eq(account.userId, userId)).limit(1);
