+++ apps/api/src/plugins/ideas/index.ts
@@ -1,7 +1,7 @@
+import { db } from "@ponti/utils";
 import { notes } from "@ponti/utils/schema";
 import { and, desc, eq } from "drizzle-orm";
 import type { FastifyInstance, FastifyPluginAsync } from "fastify";
-import { db } from "../../db";
 import type { RequestWithSession } from "../../typings";
 import { verifySession } from "../auth/utils";
