+++ apps/api/src/plugins/invites/get/index.ts
@@ -1,7 +1,7 @@
+import { db } from "@ponti/utils";
 import { list, listInvite, users } from "@ponti/utils/schema";
 import { and, asc, eq } from "drizzle-orm";
 import type { FastifyInstance } from "fastify";
-import { db } from "../../../db";
 import type { RequestWithSession } from "../../../typings";
 import { verifySession } from "../../auth/utils";
