+++ apps/api/src/services/user.service.ts
@@ -1,8 +1,8 @@
+import { db } from "@ponti/utils";
 import { list, token, users } from "@ponti/utils/schema";
 import { add } from "date-fns";
 import { eq } from "drizzle-orm";
 import { randomUUID } from "node:crypto";
-import { db } from "../db";
 
 export class UserService {
 	async createOrUpdateUser(email: string, accessToken: string) {
