+++ packages/utils/src/db/index.ts
@@ -1,7 +1,8 @@
-// biome-ignore lint/style/useNodejsImportProtocol: <explanation>
-import assert from "assert";
-import { drizzle, type NeonDatabase } from "drizzle-orm/neon-serverless";
-import * as schema from "./schema";
+import { schema } from "@ponti/utils";
+import { drizzle, type PostgresJsDatabase } from "drizzle-orm/postgres-js";
+import fastifyPlugin from "fastify-plugin";
+import assert from "node:assert";
+import postgres from "postgres";
 
 const DATABASE_URL =
 	process.env.NODE_ENV === "test"
@@ -10,6 +11,19 @@ const DATABASE_URL =
 
 assert(DATABASE_URL, "Missing DATABASE_URL");
 
-export const db: NeonDatabase<typeof schema> = drizzle(DATABASE_URL, {
+const client = postgres(DATABASE_URL);
+
+export const db: PostgresJsDatabase<typeof schema> = drizzle(client, {
 	schema,
 });
+
+export const takeOne = <T>(values: T[]): T => {
+	return values[0];
+};
+
+export const takeUniqueOrThrow = <T>(values: T[]): T => {
+	if (!Array.isArray(values)) return values;
+	if (values.length !== 1)
+		throw new Error("Found non unique or inexistent value");
+	return values[0];
+};
