+++ packages/utils/src/index.ts
@@ -7,7 +7,7 @@ export * from "./career/company.service";
 export * from "./career/job.service";
 
 // Data
-export { db } from "./db";
+export { db, takeOne, takeUniqueOrThrow } from "./db";
 export * as schema from "./db/schema";
 export { redis } from "./redis";
